+++ tests/integration_tests/test_instances.py
@@ -2,8 +2,8 @@ import os
 
 import pytest
 
+from verda import VerdaClient
 from verda.constants import Locations
-from verda.verda import VerdaClient
 
 IN_GITHUB_ACTIONS = os.getenv('GITHUB_ACTIONS') == 'true'